feat: from graph refacto code to make it more readable and add doc
